requirements.txt
discord.py
pylint
black

